moved app field to a utils file
