+++ src/input.rs
@@ -1,7 +1,7 @@
 use gtk4 as gtk;
 use gtk::gio::AppInfo;
 use gtk::{Application, ApplicationWindow, Shortcut, SearchBar, EventControllerKey};
-use crate::run::AppField;
+use crate::utils::AppField;
 
 pub fn input_handling(application: &Application, window: &ApplicationWindow) {
     // select specific app, then fetch the exec
