+++ src/main.rs
@@ -5,6 +5,7 @@ use gtk::{glib, Application};
 
 pub mod run;
 pub mod input;
+pub mod utils;
 
 const APP_ID: &str = "org.tau.Main";
