+++ src/run.rs
@@ -8,26 +8,7 @@ use gtk::{
     gio, EventControllerKey, ffi::gtk_shortcut_controller_add_shortcut,
 };
 use crate::input::{self, input_handling};
-
-#[derive(Debug, Clone)]
-pub struct AppField {
-    app_name: String,
-    exec: String,
-}
-
-impl AppField {
-    pub fn new() -> Self {
-        Self {
-            app_name: String::new(),
-            exec: String::new(),
-        }
-    }
-
-    pub fn update_fields(&mut self, app_name: String, exec: String) {
-        self.app_name = app_name;
-        self.exec = exec;
-    }
-}
+use crate::utils::AppField;
 
 pub fn draw_ui(application: &Application) {
